slice: fix Rotate for non-coprime rotations
When the rotation is not coprime with the length of the slice, there will be
multiple permutation cycles that we have to chase to cover the input.  I missed
this previously because I had copied the test from another function and then
forgot to update it to call Rotate. Brilliant.

Anyway, the tests are now fixed, and the code too.