mlink: allow cursors to move in both directions
Update the Cursor to keep track of the path from the head of the list to its
current position. This permits the cursor to move backward in the list as well
as forward, and makes it easier to detect invalidations.

Also improve the handling of cursor invalidation, and add some tests for it.