heapq: add support for intrusion
Add an Update method to the queue that allows the caller to set a callback that
will be invoked whenever the offset of an item in the queue is updated. This
allows the caller to keep track of the offsets of items in the queue as they
are moved by the addition and removal of items.