slice: rename the edit opcodes to be easier to understand
The names "insert" and "delete" imply we are modifying the inputs, which the
edit sequence does not do. Use "drop" instead of "delete", repurpose "copy" in
place of "insert", and use "emit" instead of "copy". (That is more confusing in
the moment, but it will be better when seen in isolation).