slice: deprecate Dedup, shim to the equivalent stdlib slices.Compact
There is a slight semantic change: elements beyond the end of the compacted
slice are zeroed in slices.Compact. While this is a semantic change, it should
be a strict improvement since people don't generally intend to hide old values
(and pin pointers, from the GC's POV) in the unused part of a backing array.