distinct: avoid floating-point computation in the Add step
Instead of tracking the current eviction probability as floating-point, take
advantage of the fact that it always has the form 1/2^k and use an integer
threshold instead. This lets us avoid floating-point computations both for RNG
sampling and for updates.

While here, make the Count method return unsigned instead of signed, since the
value will never be negative (even in a case of undercount).