shell: add a Reset method to the Scanner
A lot of the allocation for a Scanner is during its initialization. The Reset
method allows the caller to reuse an existing scanner with a new input reader,
keeping the expensive buffers.

Use this to pool scanners for the Split function.  On the Split benchmarks,
this helps a lot, especially for small inputs. Pooling doesn't help for
scanners managed by the caller, but Reset lets them make their own.

                        │ before   │ after
                        │ sec/op   │ sec/op    vs base
    Split/len_4-10        441.10n    72.77n    -83.50%
    Split/len_64-10        753.5n    398.9n    -47.06%
    Split/len_1024-10      6.062µ    5.822µ     -3.96%
    Split/len_16384-10     88.91µ    93.91µ     +5.61%
    Split/len_100739-10    615.8µ    652.7µ     +5.99%

                        │ before   │ after
                        │ B/op     │ B/op      vs base
    Split/len_4-10         4195.00     35.00   -99.17%
    Split/len_64-10        4256.00     96.00   -97.74%
    Split/len_1024-10      6.000Ki   1.062Ki   -82.29%
    Split/len_16384-10     22.44Ki   16.50Ki   -26.46%
    Split/len_100739-10    112.7Ki   100.8Ki   -10.59%