all: change the signature of Each to be a range function
Most of the collections defined here already have an Each method that iterates
the contents of the collection and was almost compatible with the new range
function interface added in Go 1.23.

Here, we make it fully compatible, by discarding the Boolean return value from
the Each method itself. This bumps the required Go version from 1.22 to 1.23.
Technically we could avoid that by not updating the usage within the package
itself, but as far as I am aware all the consumers of this package are using Go
1.23 already.