heapq: avoid repeated position updates during heap order fixes
Previously, when an item is moved up or down in the heap we would call an
update hook for each step of the process. This is not necessary: We need only
call it once when the value reaches its final destination.

When swapping values, only update the position of each value that is pushed out
of the way to restore heap order, and then do just one single move of the
source item at the end (if it changed).  This does not change the semantics,
but avoids unnecessary constant overhead on each update.