mnet: implement a virtual in-memory network
This package implements the net.Listener and net.Conn types in a manner
suitable for communication between goroutines in-process and for testing.
It uses net.Pipe under the covers, which exchanges data using channels, so it
can be used with the synctest package (unlike OS I/O streams, even local ones).