heapq: drop the backing array entirely in Clear
Previously Clear would simply reslice the backing array down to empty.  This
allowed the array to be reused, but at the cost of pinning any values
referenced from the unsliced portion of the array from the GC.

We could fix that problem by clearing before reslicing, but then in the case
where the caller handed us a slice via NewWithData, we'd also be clearing the
contents of that slice.

So instead, simply nil out the backing array and let subsequent operations
allocate a fresh one. This solves both problems, and is simpler, at the cost of
fresh allocations after clearing.