mnet: add the ability to emulate random port assignment (#26)
When listening on an address for a "tcp" address, if the requested address ends
in ":0", replace the literal "0" with an arbitrarily-assigned unused port
number.  This imitates how net.Listen works, though unlike the real package it
assigns ports sequentially (rather than randomly), and does not validate the
structure of the host portion of the address in that case.